mejorar la gestión del PDF clínico: se implementa un servicio para generar fichas clínicas en formato PDF y se añaden propiedades de configuración para la clínica

+++ src/main/kotlin/cl/clinipets/servicios/api/ServicioDtos.kt
@@ -9,8 +9,8 @@ import java.util.UUID
 
 data class ReglaPrecioDto(
     val id: UUID,
-    val pesoMin: BigDecimal,
-    val pesoMax: BigDecimal,
+    val pesoMin: Double,
+    val pesoMax: Double,
     val precio: Int
 )
